Added animated spritesheet for player and static spritesheet for enemies
Spritesheets have been added in the directory res/img/characters_128 for the
player and ther enemies.  Each sprite cell is 128px x 128px.  Transparent area
is colored #FF0099 (255, 000, 153).

The enemy spritesheets (apatosaurus.bmp, triceratops.bmp, tyrannosaurus.bmp, and
raptor.bmp) are 2 x 3 cells.  The first column has static "standing" images, one
each for the front, side, and back of the character.  The second column has
static "walking" images, one each for the front, side, and back of the
character.

The player spritesheet (player.bmp) is 6 x 3 cells.  The first column has static
"standing" images, one each for the front, side, and back of the player.  The
next five columns are the frames of a walking animation.  The animation order
should be 1-2-3-4-5-4-3-2 and loop.  There is an animation each for the front,
side, and back of the player.  The second column (first animation frame) can be
used as a static "walking" image, if desired.